make a lib function to get preserve drawing buffer
